More refactoring
Created folder to host all possible data fetchers
Facade remains in fetchers.py
Improved unit test
